backend/gamblegalaxy/betting/serializers.py
from rest_framework import serializers
from .models import Bet, Match

class MatchSerializer(serializers.ModelSerializer):
    class Meta:
        model = Match
        fields = '__all__'

class BetSerializer(serializers.ModelSerializer):
    class Meta:
        model = Bet
        fields = ['id', 'match', 'selected_option', 'amount']
